Added top3leaderboard.jsx
